Contact form now separate component
